server-src/ml/requirements.txt
# Neural Spaced Mastery ML Dependencies
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
psycopg2-binary==2.9.9

